+++ server/internal/models/models.go
@@ -40,7 +40,7 @@ type ConfigServer struct {
 	SubmitFlagCheckerTime uint64 `json:"submit_flag_checker_time"` // Time interval (s) to check and submit flags
 	HostFlagchecker       string `json:"host_flagchecker"`         // Address of the flagchecker server
 	TeamToken             string `json:"team_token"`               // Authentication token for team identity
-	MaxFlagBatchSize      uint16 `json:"max_flag_batch_size"`      // Max number of flags to send in a single batch
+	MaxFlagBatchSize      uint   `json:"max_flag_batch_size"`      // Max number of flags to send in a single batch
 	Protocol              string `json:"protocol"`                 // Protocol used to communicate with the flagchecker server
 }
